Moved OTR dialogs in separate package. Fixed some broken javadoc
