+++ src/net/java/sip/communicator/plugin/otr/OtrConfigurationPanel.java
@@ -16,6 +16,7 @@ import javax.swing.border.*;
 import net.java.otr4j.*;
 import net.java.sip.communicator.service.protocol.*;
 import net.java.sip.communicator.plugin.desktoputil.*;
+import net.java.sip.communicator.plugin.otr.authdialog.*;
 
 /**
  * A special {@link Panel} that manages the OTR configuration.
