+++ src/net/java/sip/communicator/plugin/otr/ScOtrEngineImpl.java
@@ -6,20 +6,14 @@
  */
 package net.java.sip.communicator.plugin.otr;
 
-import java.awt.*;
-import java.awt.event.*;
 import java.net.*;
 import java.security.*;
 import java.util.*;
-import java.util.List;
 import java.util.concurrent.*;
-
-import javax.swing.*;
-import javax.swing.plaf.basic.*;
 
 import net.java.otr4j.*;
 import net.java.otr4j.session.*;
-import net.java.sip.communicator.plugin.desktoputil.*;
+import net.java.sip.communicator.plugin.otr.authdialog.*;
 import net.java.sip.communicator.service.browserlauncher.*;
 import net.java.sip.communicator.service.contactlist.*;
 import net.java.sip.communicator.service.gui.*;
@@ -881,314 +875,5 @@ public class ScOtrEngineImpl
             showError(session.getSessionID(), e.getMessage());
         }
         
-    }
-
-    /**
-     * The dialog that pops up when SMP negotiation starts.
-     * It contains a progress bar that indicates the status of the SMP
-     * authentication process.
-     */
-    @SuppressWarnings("serial")
-    private class SmpProgressDialog
-        extends SIPCommDialog
-    {
-        private final JProgressBar progressBar = new JProgressBar(0, 100);
-
-        private final Color successColor = new Color(86, 140, 2);
-
-        private final Color failColor = new Color(204, 0, 0);
-
-        private final JLabel iconLabel = new JLabel();
-
-        /**
-         * Instantiates SmpProgressDialog.
-         * 
-         * @param contact The contact that this dialog is associated with.
-         */
-        public SmpProgressDialog(Contact contact)
-        {
-            setTitle(
-                OtrActivator.resourceService.getI18NString(
-                    "plugin.otr.smpprogressdialog.TITLE"));
-
-            JPanel mainPanel = new TransparentPanel();
-            mainPanel.setLayout(new BoxLayout(mainPanel, BoxLayout.Y_AXIS));
-            mainPanel.setBorder(
-                BorderFactory.createEmptyBorder(10, 10, 10, 10));
-            mainPanel.setPreferredSize(new Dimension(300, 70));
-
-            String authFromText =
-                String.format(
-                    OtrActivator.resourceService
-                        .getI18NString(
-                            "plugin.otr.authbuddydialog.AUTHENTICATION_FROM",
-                            new String[] {contact.getDisplayName()}));
-
-            JPanel labelsPanel = new TransparentPanel();
-            labelsPanel.setLayout(new BoxLayout(labelsPanel, BoxLayout.X_AXIS));
-
-            labelsPanel.add(iconLabel);
-            labelsPanel.add(Box.createRigidArea(new Dimension(5,0)));
-            labelsPanel.add(new JLabel(authFromText));
-
-            mainPanel.add(labelsPanel);
-            mainPanel.add(progressBar);
-
-            init();
-
-            this.getContentPane().add(mainPanel);
-            this.pack();
-        }
-
-        /**
-         * Initializes the progress bar and sets it's progression to 1/3.
-         */
-        public void init()
-        {
-            progressBar.setUI(new BasicProgressBarUI() {
-                private Rectangle r = new Rectangle();
-
-                @Override
-                protected void paintIndeterminate(Graphics g, JComponent c) {
-                    Graphics2D g2d = (Graphics2D) g;
-                    g2d.setRenderingHint(
-                        RenderingHints.KEY_ANTIALIASING,
-                        RenderingHints.VALUE_ANTIALIAS_ON);
-                    r = getBox(r);
-                    g.setColor(progressBar.getForeground());
-                    g.fillOval(r.x, r.y, r.width, r.height);
-                }
-            });
-            progressBar.setValue(33);
-            progressBar.setForeground(successColor);
-            progressBar.setStringPainted(false);
-            iconLabel.setIcon(
-                OtrActivator.resourceService.getImage(
-                    "plugin.otr.ENCRYPTED_UNVERIFIED_ICON_22x22"));
-        }
-
-        /**
-         * Sets the progress bar to 2/3 of completion.
-         */
-        public void incrementProgress()
-        {
-            progressBar.setValue(66);
-        }
-
-        /**
-         * Sets the progress bar to green.
-         */
-        public void setProgressSuccess()
-        {
-            progressBar.setValue(100);
-            progressBar.setForeground(successColor);
-            progressBar.setStringPainted(true);
-            progressBar.setString(
-                OtrActivator.resourceService
-                    .getI18NString(
-                        "plugin.otr.smpprogressdialog.AUTHENTICATION_SUCCESS"));
-            iconLabel.setIcon(
-                OtrActivator.resourceService.getImage(
-                    "plugin.otr.ENCRYPTED_ICON_22x22"));
-        }
-
-        /**
-         * Sets the progress bar to red.
-         */
-        public void setProgressFail()
-        {
-            progressBar.setValue(100);
-            progressBar.setForeground(failColor);
-            progressBar.setStringPainted(true);
-            progressBar.setString(
-                OtrActivator.resourceService
-                    .getI18NString(
-                        "plugin.otr.smpprogressdialog.AUTHENTICATION_FAIL"));
-        }
-    }
-
-    /**
-     * The dialog that pops up when the remote party send us SMP
-     * request. It contains detailed information for the user about
-     * the authentication process and allows him to authenticate.
-     *
-     */
-    @SuppressWarnings("serial")
-    private class SmpAuthenticateBuddyDialog
-        extends SIPCommDialog
-    {
-        private final Contact contact;
-
-        private final String question;
-
-        SmpAuthenticateBuddyDialog(Contact contact, String question)
-        {
-            this.contact = contact;
-            this.question = question;
-            initComponents();
-        }
-
-        private void initComponents()
-        {
-            this.setTitle(
-                OtrActivator.resourceService
-                    .getI18NString(
-                        "plugin.otr.authbuddydialog.TITLE"));
-
-            // The main panel that contains all components.
-            JPanel mainPanel = new TransparentPanel();
-            mainPanel.setLayout(new BoxLayout(mainPanel, BoxLayout.Y_AXIS));
-            mainPanel.setBorder(
-                BorderFactory.createEmptyBorder(10, 10, 10, 10));
-            mainPanel.setPreferredSize(new Dimension(300, 350));
-
-            // Add "authentication from contact" to the main panel.
-            JTextArea authenticationFrom = new CustomTextArea();
-            Font newFont =
-                new Font(
-                    UIManager.getDefaults().getFont("TextArea.font").
-                        getFontName()
-                    , Font.BOLD
-                    , 14);
-            authenticationFrom.setFont(newFont);
-            String authFromText =
-                String.format(
-                    OtrActivator.resourceService
-                        .getI18NString(
-                            "plugin.otr.authbuddydialog.AUTHENTICATION_FROM",
-                            new String[] {contact.getDisplayName()}));
-            authenticationFrom.setText(authFromText);
-            mainPanel.add(authenticationFrom);
-
-            // Add "general info" text to the main panel.
-            JTextArea generalInfo = new CustomTextArea();
-            generalInfo.setText(OtrActivator.resourceService
-                .getI18NString(
-                    "plugin.otr.authbuddydialog.AUTHENTICATION_INFO"));
-            mainPanel.add(generalInfo);
-
-            // Add "authentication-by-secret" info text to the main panel.
-            JTextArea authBySecretInfo = new CustomTextArea();
-            newFont =
-                new Font(
-                    UIManager.getDefaults().getFont("TextArea.font").
-                        getFontName()
-                    , Font.ITALIC
-                    , 10);
-            authBySecretInfo.setText(OtrActivator.resourceService
-                .getI18NString(
-                    "plugin.otr.authbuddydialog.AUTH_BY_SECRET_INFO_RESPOND"));
-            authBySecretInfo.setFont(newFont);
-            mainPanel.add(authBySecretInfo);
-
-            // Create a panel to add question/answer related components
-            JPanel questionAnswerPanel = new JPanel(new GridBagLayout());
-            questionAnswerPanel.setBorder(BorderFactory.createEtchedBorder());
-
-            GridBagConstraints c = new GridBagConstraints();
-            c.gridx = 0;
-            c.gridy = 0;
-            c.fill = GridBagConstraints.HORIZONTAL;
-            c.insets = new Insets(5, 5, 0, 5);
-            c.weightx = 0;
-
-            // Add question label.
-            JLabel questionLabel =
-                new JLabel(
-                    OtrActivator.resourceService
-                        .getI18NString(
-                            "plugin.otr.authbuddydialog.QUESTION_RESPOND"));
-            questionAnswerPanel.add(questionLabel, c);
-
-            // Add the question.
-            c.insets = new Insets(0, 5, 5, 5);
-            c.gridy = 1;
-            JTextArea questionArea = 
-                new CustomTextArea();
-            newFont =
-                new Font(
-                    UIManager.getDefaults().getFont("TextArea.font").
-                        getFontName()
-                    , Font.BOLD
-                    , UIManager.getDefaults().getFont("TextArea.font")
-                        .getSize());
-            questionArea.setFont(newFont);
-            questionArea.setText(question);
-            questionAnswerPanel.add(questionArea, c);
-
-            // Add answer label.
-            c.insets = new Insets(5, 5, 5, 5);
-            c.gridy = 2;
-            JLabel answerLabel =
-                new JLabel(OtrActivator.resourceService
-                    .getI18NString("plugin.otr.authbuddydialog.ANSWER"));
-            questionAnswerPanel.add(answerLabel, c);
-
-            // Add the answer text field.
-            c.gridy = 3;
-            final JTextField answerTextBox = new JTextField();
-            questionAnswerPanel.add(answerTextBox, c);
-
-            // Add the question/answer panel to the main panel.
-            mainPanel.add(questionAnswerPanel);
-
-            // Buttons panel.
-            JPanel buttonPanel = new TransparentPanel(new GridBagLayout());
-            buttonPanel.setBorder(BorderFactory.createEmptyBorder(0, 5, 5, 5));
-
-            JButton helpButton =
-                new JButton(OtrActivator.resourceService
-                    .getI18NString("plugin.otr.authbuddydialog.HELP"));
-            helpButton.addActionListener(new ActionListener()
-            {
-                public void actionPerformed(ActionEvent arg0)
-                {
-                    OtrActivator.scOtrEngine.launchHelp();
-                }
-            });
-
-            c.gridwidth = 1;
-            c.gridy = 0;
-            c.gridx = 0;
-            c.weightx = 0;
-            c.insets = new Insets(5, 5, 5, 20);
-            buttonPanel.add(helpButton, c);
-
-            JButton cancelButton =
-                new JButton(OtrActivator.resourceService
-                    .getI18NString("plugin.otr.authbuddydialog.CANCEL"));
-            cancelButton.addActionListener(new ActionListener()
-            {
-                public void actionPerformed(ActionEvent e)
-                {
-                    ScOtrEngineImpl.this.abortSmp(contact);
-                    SmpAuthenticateBuddyDialog.this.dispose();
-                }
-            });
-            c.insets = new Insets(5, 5, 5, 5);
-            c.gridx = 1;
-            buttonPanel.add(cancelButton, c);
-
-            c.gridx = 2;
-            JButton authenticateButton =
-                new JButton(OtrActivator.resourceService
-                    .getI18NString(
-                        "plugin.otr.authbuddydialog.AUTHENTICATE_BUDDY"));
-            authenticateButton.addActionListener(new ActionListener()
-            {
-                public void actionPerformed(ActionEvent e)
-                {
-                    ScOtrEngineImpl.this.respondSmp(
-                        contact, question, answerTextBox.getText());
-                    SmpAuthenticateBuddyDialog.this.dispose();
-                }
-            });
-
-            buttonPanel.add(authenticateButton, c);
-
-            this.getContentPane().add(mainPanel, BorderLayout.NORTH);
-            this.getContentPane().add(buttonPanel, BorderLayout.SOUTH);
-            this.pack();
-        }
     }
 }
