+++ src/net/java/sip/communicator/plugin/otr/authdialog/KnownFingerprintsPanel.java
@@ -4,7 +4,7 @@
  * Distributable under LGPL license.
  * See terms of license at gnu.org.
  */
-package net.java.sip.communicator.plugin.otr;
+package net.java.sip.communicator.plugin.otr.authdialog;
 
 import java.awt.*;
 import java.awt.event.*;
@@ -15,6 +15,7 @@ import javax.swing.event.*;
 
 import net.java.sip.communicator.service.protocol.*;
 import net.java.sip.communicator.plugin.desktoputil.*;
+import net.java.sip.communicator.plugin.otr.*;
 
 /**
  * @author @George Politis
