+++ src/net/java/sip/communicator/plugin/otr/authdialog/OtrBuddyAuthenticationDialog.java
@@ -4,7 +4,7 @@
  * Distributable under LGPL license.
  * See terms of license at gnu.org.
  */
-package net.java.sip.communicator.plugin.otr;
+package net.java.sip.communicator.plugin.otr.authdialog;
 
 import java.awt.*;
 import java.awt.event.*;
@@ -12,7 +12,8 @@ import java.awt.event.*;
 import javax.swing.*;
 
 import net.java.sip.communicator.plugin.desktoputil.*;
-import net.java.sip.communicator.plugin.otr.FingerprintAuthenticationPanel.ActionComboBoxItem;
+import net.java.sip.communicator.plugin.otr.*;
+import net.java.sip.communicator.plugin.otr.authdialog.FingerprintAuthenticationPanel.*;
 import net.java.sip.communicator.service.protocol.*;
 
 /**
